chore: moved test util
